refactor(rda_json): move service location
This change is a pure refactor. No behavioural changes are expected.

Move `spec/services/plan_snapshot_rda_json.rb` to `app/services/plan_snapshots/rda_json.rb`
- Move corresponding spec file as well
- Update caller in `PlanSnapshot` model to correspond with the changed file location.
